Update raster manager and start implementing raster value extractor

+++ niamoto/raster/raster_manager.py
@@ -88,8 +88,8 @@ class RasterManager:
             connection.execute(ins)
 
     @classmethod
-    def update_raster(cls, raster_file_path, name, tile_dimension=None,
-                      srid=None):
+    def update_raster(cls, raster_file_path, name, new_name=None,
+                      tile_dimension=None, srid=None):
         """
         Update an existing raster in database and register it the Niamoto
         raster registry. Uses raster2pgsql command. The raster is cut in
@@ -97,6 +97,7 @@ class RasterManager:
         are stored
         :param raster_file_path: The path to the raster file.
         :param name: The name of the raster.
+        :param new_name: The new name of the raster (not changed if None).
         :param tile_dimension: The tile dimension (width, height), if None,
             tile dimension will be chosen automatically by PostGIS.
         :param srid: SRID to assign to stored raster. If None, use raster's
@@ -113,11 +114,16 @@ class RasterManager:
             dim = "{}x{}".format(tile_dimension[0], tile_dimension[1])
         else:
             dim = 'auto'
-        tb = "{}.{}".format(settings.NIAMOTO_RASTER_SCHEMA, name)
         os.environ["PGPASSWORD"] = settings.NIAMOTO_DATABASE["PASSWORD"]
         dev_null = open(os.devnull, 'w')  # Force quiet
+        d = "-d"
+        if new_name is None:
+            new_name = name
+        else:
+            d = "-c"
+        tb = "{}.{}".format(settings.NIAMOTO_RASTER_SCHEMA, new_name)
         p1 = subprocess.Popen([
-            "raster2pgsql", "-d", '-t', dim, '-I', raster_file_path, tb,
+            "raster2pgsql", d, '-t', dim, '-I', raster_file_path, tb,
         ], stdout=subprocess.PIPE, stderr=dev_null)
         dev_null.close()
         p2 = subprocess.call([
@@ -134,28 +140,36 @@ class RasterManager:
         if p2 != 0:
             raise RuntimeError("raster import failed.")
         upd = niamoto_db_meta.raster_registry.update().values({
+            'name': new_name,
             'srid': srid,
             'date_update': datetime.now(),
         }).where(niamoto_db_meta.raster_registry.c.name == name)
         with Connector.get_connection() as connection:
             connection.execute(upd)
-
-    @classmethod
-    def delete_raster(cls, name):
-        """
-        Delete an existing raster.
-        :param name: The name of the raster.
-        """
-        cls.assert_raster_exists(name)
-        with Connector.get_connection() as connection:
-            with connection.begin():
+            if new_name != name:
                 connection.execute("DROP TABLE IF EXISTS {};".format(
                     "{}.{}".format(settings.NIAMOTO_RASTER_SCHEMA, name)
                 ))
-                del_stmt = niamoto_db_meta.raster_registry.delete().where(
-                    niamoto_db_meta.raster_registry.c.name == name
-                )
-                connection.execute(del_stmt)
+
+    @classmethod
+    def delete_raster(cls, name, connection=None):
+        """
+        Delete an existing raster.
+        :param name: The name of the raster.
+        :param connection: If provided, use an existing connection.
+        """
+        cls.assert_raster_exists(name)
+        if connection is None:
+            connection = Connector.get_engine().connect()
+        with connection.begin():
+            connection.execute("DROP TABLE IF EXISTS {};".format(
+                "{}.{}".format(settings.NIAMOTO_RASTER_SCHEMA, name)
+            ))
+            del_stmt = niamoto_db_meta.raster_registry.delete().where(
+                niamoto_db_meta.raster_registry.c.name == name
+            )
+            connection.execute(del_stmt)
+        connection.close()
 
     @classmethod
     def get_raster_srid(cls, raster_file_path):
